Semantic editor displays treeview objects
Semantic editor now displays treeview stuff rather than the canvas stuff. also updated the canvas items so that they all have unique UUID's and store their original UUID within themselves for use in functions. this was bec ause a lot of the functionality in this program hinges off of the fact that the canvas objects need to have unique identifiers and a lot of functions used the semantic UUID